refactor: extract chat logic into ChatViewModel
