+++ app/src/main/java/com/codenames/frontend/viewmodel/GameViewModel.kt
@@ -3,15 +3,11 @@ package com.codenames.frontend.viewmodel
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.codenames.frontend.data.model.ChatLists
 import com.codenames.frontend.data.model.GameState
 import com.codenames.frontend.data.model.enums.CardType
-import com.codenames.frontend.data.model.enums.ChatTab
 import com.codenames.frontend.data.model.enums.ConnectionState
-import com.codenames.frontend.data.model.enums.Role
 import com.codenames.frontend.data.model.enums.Team
 import com.codenames.frontend.data.model.toGameState
-import com.codenames.frontend.data.repository.ChatRepository
 import com.codenames.frontend.data.repository.GameRepository
 import com.codenames.frontend.network.dto.GameMessage
 import com.codenames.frontend.network.websocket.GameWebSocketHandler
@@ -30,17 +26,12 @@ class GameViewModel
     @Inject
     constructor(
         private val client: GameWebSocketHandler,
-        private val chatRepository: ChatRepository,
         private val gameRepository: GameRepository,
     ) : ViewModel() {
         private var job: Job? = null
 
         private val _uiState = MutableStateFlow(GameState())
         val uiState: StateFlow<GameState> = _uiState
-
-        private val _chatState = MutableStateFlow(ChatLists())
-        val chatState: StateFlow<ChatLists> = _chatState
-
         private val _connectionState = MutableStateFlow<ConnectionState>(ConnectionState.IDLE)
         val connectionState: StateFlow<ConnectionState> = _connectionState
 
@@ -72,32 +63,6 @@ class GameViewModel
 
                         Log.d("GameViewModel", "Subscribed to Lobby")
 
-                        launch {
-                            chatRepository.observeChat("/topic/chat/$lobbyCode", username).collect { msg ->
-                                _chatState.update { currentState ->
-                                    currentState.copy(lobbyMessages = currentState.lobbyMessages + msg)
-                                }
-                            }
-                        }
-
-                        launch {
-                            chatRepository.observeChat("/topic/chat/$lobbyCode/$team", username).collect { msg ->
-                                _chatState.update { currentState ->
-                                    currentState.copy(teamMessages = currentState.teamMessages + msg)
-                                }
-                            }
-                        }
-
-                        if (role == Role.OPERATIVE.name) {
-                            launch {
-                                chatRepository.observeChat("/topic/chat/$lobbyCode/$team/operative", username).collect { msg ->
-                                    _chatState.update { currentState ->
-                                        currentState.copy(operativeMessages = currentState.operativeMessages + msg)
-                                    }
-                                }
-                            }
-                        }
-
                         if (isHost) {
                             delay(2000)
                             sendGameStart(lobbyCode)
@@ -116,39 +81,6 @@ class GameViewModel
                 gameRepository.startGame(lobbyCode)
             }
         }
-
-        fun sendLobbyMessage(
-            lobbyCode: String,
-            username: String,
-            content: String,
-        ) {
-            viewModelScope.launch {
-                chatRepository.sendMessage("/app/chat/$lobbyCode", username, content)
-            }
-        }
-
-        fun sendTeamMessage(
-            lobbyCode: String,
-            team: String,
-            username: String,
-            content: String,
-        ) {
-            viewModelScope.launch {
-                chatRepository.sendMessage("/app/chat/$lobbyCode/$team", username, content)
-            }
-        }
-
-        fun sendOperativeMessage(
-            lobbyCode: String,
-            team: String,
-            username: String,
-            content: String,
-        ) {
-            viewModelScope.launch {
-                chatRepository.sendMessage("/app/chat/$lobbyCode/$team/operative", username, content)
-            }
-        }
-
         fun submitClue(
             lobbyCode: String,
             word: String,
@@ -170,53 +102,12 @@ class GameViewModel
                 }
             }
         }
-
-        fun sendChatMessage(
-            tab: ChatTab,
-            lobbyCode: String,
-            username: String,
-            team: Team?,
-            content: String,
-            availableChatTabs: List<ChatTab>,
-        ) {
-            if (lobbyCode.isBlank()) {
-                return
-            }
-
-            when (tab) {
-                ChatTab.GLOBAL ->
-                    sendLobbyMessage(
-                        lobbyCode = lobbyCode,
-                        username = username,
-                        content = content,
-                    )
-
-                ChatTab.TEAM ->
-                    if (team != null) {
-                        sendTeamMessage(
-                            lobbyCode = lobbyCode,
-                            team = team.name,
-                            username = username,
-                            content = content,
-                        )
-                    }
-
-                ChatTab.OPERATIVES ->
-                    if (team != null && ChatTab.OPERATIVES in availableChatTabs) {
-                        sendOperativeMessage(
-                            lobbyCode = lobbyCode,
-                            team = team.name,
-                            username = username,
-                            content = content,
-                        )
-                    }
-            }
-        }
-
         fun handleMessage(message: GameMessage) {
             val state = message.toGameState()
-            _uiState.update {
-                state
+            _uiState.update { current ->
+                state.copy(
+                    chatLists = current.chatLists
+                )
             }
             Log.d("GameViewModel", "Updated game state: $state")
         }
